requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.2
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
dj-database-url==0.5.0
Django==3.2.5
django-enumchoicefield==2.0.0
django-environ==0.4.5
django-heroku==0.3.1
django-rest-enumfield==0.2.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
openapi-codec==1.3.2
psycopg2==2.9.1
PyJWT==2.1.0
python-dotenv==0.19.0
python-http-client==3.3.2
pytz==2021.1
requests==2.26.0
sendgrid==6.8.1
simplejson==3.17.3
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.2.0
